refactor: 회원 가입 API 패키지 변경 #1
- member -> auth 로 변경
