+++ internal/auth/dto.go
@@ -1,0 +1,8 @@
+package auth
+
+type SignupRequest struct {
+	Name        string `json:"name" binding:"required,min=1,max=20"`
+	Email       string `json:"email" binding:"required,email,max=50"`
+	PhoneNumber string `json:"phoneNumber" binding:"required,phone"`
+	Password    string `json:"password" binding:"required,min=8,max=15"`
+}
